Rename headers to descriptive names
