+++ include/vsuite/fixed.h
@@ -1,5 +1,5 @@
-#ifndef VSUITE_VF_H
-#define VSUITE_VF_H
+#ifndef VSUITE_FIXED_H
+#define VSUITE_FIXED_H
 
 #include <string.h>
 #include <stddef.h>
@@ -16,4 +16,15 @@
         }                                                                     \
     } while (0)
 
-#endif /* VSUITE_VF_H */
+/* Copy from Fixed VARCHAR to Fixed C String */
+#define fv_copy(cdst, vsrc)                                                   \
+    do {                                                                      \
+        if ((vsrc).len < sizeof(cdst)) {                                      \
+            memcpy(cdst, V_BUF(vsrc), (vsrc).len);                            \
+            cdst[(vsrc).len] = '\0';                                          \
+        } else if (sizeof(cdst) > 0) {                                        \
+            cdst[0] = '\0';                                                   \
+        }                                                                     \
+    } while (0)
+
+#endif /* VSUITE_FIXED_H */
