+++ include/vsuite/varchar.h
